requirements.txt
torch>=2.5.0
torchvision>=0.20.0
torchaudio>=2.5.0
gradio>=4.0.0
pandas>=2.0.0
numpy>=1.24.0
nltk>=3.8.0
pillow>=10.0.0
matplotlib>=3.7.0
tqdm>=4.65.0
scikit-learn>=1.3.0
kaggle>=1.5.16
